pr (#36)
* Develop juan (#33)

* agrego algunos jdbc

* implementaciones en itemMenuJDBC y correcciones varias

* traigo cambios de master (#31)

* implementaciones en ItemMenuJDBC y cambios menores (#29)

* agrego algunos jdbc

* implementaciones en itemMenuJDBC y correcciones varias

* Creamos los primeros test unitarios de ClientController (#30)


---------


* some more jdbc

---------


* Logica de agregar items a carrito, falta la visual. (#34)

* New Frame ModificarCliente

* Modificar Cliente

* Modificar Cliente

---------


* Develop lucho (#35)

* Creamos los primeros test unitarios de ClientController

* Creamos el Test para verificar que ningun ID este duplicado ClienteJDBCTest

* Creamos el Test para verificar que ningun ID este duplicado ClienteJDBCTest

* Se testea la consistencia de ClienteJDBC.

* Se testea la consistencia de ClienteJDBC y se acomodan errores de consistencia.

---------


---------

+++ src/main/java/nc/dao/jdbc/ClienteJDBC.java
@@ -44,7 +44,7 @@ public class ClienteJDBC implements ClienteDAO{
         String query = "INSERT INTO cliente (nombre, cuit, email, direccion, cx, cy) VALUES (?, ?, ?, ?, ?, ?)";
         try (PreparedStatement ps = conn.prepareStatement(query)) {
             ps.setString(1, cliente.getNombre());
-            ps.setInt(2, cliente.getCuit());
+            ps.setLong(2, cliente.getCuit());
             ps.setString(3, cliente.getEmail());
             ps.setString(4, cliente.getDireccion());
             ps.setDouble(5, cliente.getCoordenada().getLatitude());
@@ -60,7 +60,7 @@ public class ClienteJDBC implements ClienteDAO{
         String query = "UPDATE cliente SET nombre = ?, cuit = ?, email = ?, direccion = ?, cx = ?, cy = ? WHERE id = ?";
         try (PreparedStatement ps = conn.prepareStatement(query)) {
             ps.setString(1, cliente.getNombre());
-            ps.setInt(2, cliente.getCuit());
+            ps.setLong(2, cliente.getCuit());
             ps.setString(3, cliente.getEmail());
             ps.setString(4, cliente.getDireccion());
             ps.setDouble(5, cliente.getCoordenada().getLatitude());
